% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/click.R
\name{as_pixeltrix}
\alias{as_pixeltrix}
\alias{is_pixeltrix}
\title{Coerce to a 'pixeltrix' Object}
\usage{
as_pixeltrix(m)

is_pixeltrix(m)
}
\arguments{
\item{m}{A matrix of integers to coerce.}
}
\value{
\code{as_pixeltrix} returns an object of class 'pixeltrix' if
possible. \code{is_pixeltrix} returns \code{TRUE} if the object has class
'pixeltrix', otherwise \code{FALSE}. If coerced, a named character-vector
attribute, 'colours', is also added, which maps the matrix values to a
gradated palette of greys in hex form.
}
\description{
Functions to check if an object is of 'pixeltrix' class, or coerce to it if
possible.
}
\details{
To be successfully coerced, \code{m} must be a matrix composed only of
integers.
}
\examples{
m <- matrix(c(0L, 1L, 1L, 0L), 2, 2)
is_pixeltrix(m)
m

m <- as_pixeltrix(m)
is_pixeltrix(m)
m

}
