% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blue}
\alias{blue}
\title{Example Sprite Matrix: Pokemon}
\format{
A matrix with 16 rows and 14 columns, taking values of 0 to 2.
}
\source{
Hand-copied from Pokemon Blue (1996), The Pokemon Company.
}
\usage{
blue
}
\description{
A matrix output from \code{\link{click_pixels}} that represents the pixels of
a player sprite from Pokemon Blue.
}
\examples{
\dontrun{
# Plot the matrix as an image
draw_pixels(
  m = pixeltrix::blue,
  colours = c("white", "#879afb", "gray20")
)
}

}
\keyword{datasets}
