% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_pixels}
\alias{draw_pixels}
\title{Draw 'Pixel' Matrix in a Plot Window}
\usage{
draw_pixels(m, colours = NULL, ...)
}
\arguments{
\item{m}{A matrix of integers. The maximum value is assumed to be the number
of pixel states desired. Override by supplying a 'n_states' value larger
than the maximum in the matrix.}

\item{colours}{A character vector of named colours. One for each value in m,
the provided matrix. The order you provide the colours matches the order
of the values in the matrix, from 0 to n. Defaults to NULL, which means
random colours will be selected on your behalf.}

\item{...}{Additional graphical parameters passed to
\code{\link[graphics]{image}} and thus \code{\link[graphics]{plot}}.}
}
\value{
Nothing.
}
\description{
Opens a plot window and draws a provided matrix. Designed for use with the
output from \code{\link{click_pixels}} (or \code{\link{edit_pixels}}).
}
\examples{
\dontrun{
  my_matrix <- click_pixels(n_states = 3)
  draw_pixels(my_matrix, c("black", "#0000FF", "green"))  # one colour per state
}
}
