% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{frame_pixels}
\alias{frame_pixels}
\title{Create Frames of a Pixel Animation}
\usage{
frame_pixels(
  n_rows = 8L,
  n_cols = 8L,
  n_states = 2L,
  colours = NULL,
  grid = TRUE
)
}
\arguments{
\item{n_rows}{Integer. The number of pixels high that the plot should be.
Numeric values are coerced to integer.}

\item{n_cols}{Integer. The number of pixels wide that the plot should be.
Numeric values are coerced to integer.}

\item{n_states}{Integer. The number of states that a pixel can be. Click a
pixel to cycle through the states. Numeric values are coerced to integer.
See details.}

\item{colours}{Character vector. As many named/hex colours as n_state. Each
click in the interactive plot will cycle a pixel through these colours.
Defaults to NULL, which results in an attempt to find and use the
'colours' attribute (named character vector) of the input matrix, 'm'.
This attribute is added by default to matrices created with
\code{\link{click_pixels}} (recommended).}

\item{grid}{Logical. Should a boundary line be placed around the pixels to
make them easier to differentiate between them? Defaults to TRUE.}
}
\value{
A list of matrices.
}
\description{
Opens a new interactive plotting canvas with a grid of clickable squares
('pixels'). When finished, the user is prompted to provide another.
}
\details{
Click the pixels in the plotting window repeatedly to cycle through
a number of 'states'. Successive clicks increase the state value by 1
(wrapping back to 0, the default when the canvas is first plotted) and
make the pixel darker grey in colour. Press the ESCAPE key to exit the
mode. You'll be prompted interactively in the console to add a new frame
or not. If you refuse a new frame, you'll be returned a list of matrices,
each of which contain the state values of each pixel for each frame of
your animation.

If your editor opens a separate graphics window (i.e. not RStudio),
each click may result in a brief flash as the image refreshes, while a
resized window may return to its original dimensions. You may also hear a
bell sound on click, which you can disable by setting
\code{options(locatorBell = FALSE)}.
}
\examples{
\dontrun{
    # Begin interactive sequence to create animation frames
    my_frames <- frame_pixels(
      n_rows   = 16L,
      n_cols   = 16L,
      n_states = 3L,
      colours  = c("grey25", "green", "#0000FF")
    )
}
}
