% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{gif_pixels}
\alias{gif_pixels}
\title{Write Frames of a Pixel Animation to GIF}
\usage{
gif_pixels(frames, colours = NULL, file, ...)
}
\arguments{
\item{frames}{A list of matrices. preferably produced by \code{\link{frame_pixels}}.
Each matrix in the list is a frame of the final animation.}

\item{colours}{A character vector of named colours. One for each unique value
in the matrices of the 'frames' list. The order you provide the colours
matches the sorted order of the values in the matrix. Defaults to NULL,
which means colours will be extracted from the 'colours' attribute of the
matrices (which is added to the object when using the recommended
\code{\link{frame_pixels}} function), otherwise a gradated palette of
greys will be selected.}

\item{file}{A filepath expressed as a character vector, ending with
extension '.gif'. This is where the output animation will be written.}

\item{...}{Parameters to pass to \code{\link[gifski]{save_gif}} (you must
have 'gifski' installed).}
}
\value{
Nothing.
}
\description{
Plots (using \code{\link{draw_pixels}}) each matrix from a list (ideally
created using \code{\link{frame_pixels}}) and writes it to an animated GIF.
Requires the 'gifski' package.
}
\examples{
\dontrun{
# Begin interactive sequence to create animation frames
my_frames <- frame_pixels(
  n_rows   = 16L,
  n_cols   = 16L,
  n_states = 3L,
  colours  = c("grey25", "green", "#0000FF")
)

# Write list of matrices to gif (requires 'gifski' installation)
gif_pixels(
  frames = my_frames,
  file = "example.gif",  # location to write GIF
  delay = 0.1            # passed to gifski::save_gif()
)}
}
