% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{logo}
\alias{logo}
\title{Example Sprite Matrix: 'pixeltrix' Package Logo Text}
\format{
A 'pixeltrix'-class matrix with 10 rows and 21 columns, taking
integers from 0 to 2.
}
\usage{
logo
}
\description{
A matrix output from \code{\link{click_pixels}} that represents the
text of the pixeltrix package logo, which is the word 'pixeltrix' written
in a font with letters that are 3 by 3 pixels.
}
\examples{
\dontrun{
# Plot the matrix as an image
draw_pixels(m = pixeltrix::logo)}

}
\keyword{datasets}
