% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mario}
\alias{mario}
\title{Example Sprite Matrix: Mario}
\format{
A list of matrices, each with 16 rows and 16 columns, taking values
of 0 to 3.
}
\source{
Super Mario Bros (1983), Nintendo Entertainment System, Nintendo.
}
\usage{
mario
}
\description{
A list of matrices output from \code{\link{frame_pixels}} that represent the
pixels of the Mario sprite from Super Mario Bros. Each matrix is a step in
the walk cycle.
}
\examples{
\dontrun{
# Write the list of matrices to a gif
gif_pixels(
  frames = pixeltrix::mario,
  colours = c("white", "#FDA428", "#FC0D1B", "#A32B2E"),
  file = "mario.gif",
  delay = 0.15
)
}

}
\keyword{datasets}
