% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mario_frames}
\alias{mario_frames}
\title{Example Sprite Matrix: Mario}
\format{
A list of 'pixeltrix'-class matrices, each with 16 rows and 16
columns, taking integers from 0 to 2.
}
\source{
Super Mario Bros (1983), Nintendo.
}
\usage{
mario_frames
}
\description{
A list of matrices output from \code{\link{frame_pixels}} that represent the
pixels of the Mario sprite from Super Mario Bros. Each matrix is a step in
the walk cycle.
}
\examples{
\dontrun{
# Write the list of matrices to a gif
gif_pixels(
  frames = pixeltrix::mario_frames,
  file = "mario.gif",
  delay = 0.15  # passed to gifski::save_gif()
)}

}
\keyword{datasets}
