% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{pixeltrix}
\alias{pixeltrix}
\title{A Shiny App to Create Matrices from Users' Pixel Art}
\source{
\href{https://shiny.rstudio.com/gallery/image-interaction-basic.html}{RStudio template}
}
\usage{
pixeltrix()
}
\value{
Nothing.
}
\description{
A Shiny App to Create Matrices from Users' Pixel Art
}
\examples{
\dontrun{pixeltrix()}
}
