% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pkmn_sprite}
\alias{pkmn_sprite}
\title{Example Sprite Matrix: Pokémon}
\format{
A matrix with 16 rows and 14 columns, taking integers from 0 to 2.
}
\source{
Hand-copied from Pokemon (1996), The Pokemon Company.
}
\usage{
pkmn_sprite
}
\description{
A matrix output from \code{\link{click_pixels}} that represents the pixels of
a player sprite from Pokémon.
}
\examples{
\dontrun{
    # Plot the matrix as an image
    draw_pixels(
      m = pixeltrix::pkmn_sprite,
      colours = c("#9bbc0f", "#8bac0f", "#306230")
    )
}

}
\keyword{datasets}
