% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locator.R
\name{click_pixels}
\alias{click_pixels}
\title{Create Pixel Art, Get a Matrix}
\usage{
click_pixels(n_rows = 8L, n_cols = 16L)
}
\arguments{
\item{n_rows}{Integer. The number of 'pixels' high that the plot should be.}

\item{n_cols}{Integer. The number of 'pixels' wide that the plot should be.}
}
\value{
A matrix.
}
\description{
A plotting canvas will be opened with the provided dimensions with a grid of
clickable squares. These can be clicked between 'on' and 'off' states.
Press ESCAPE to exit the interactive mode and be returned a matrix with the
updated values.
}
\details{
Numeric inputs will be converted to integers.
}
\examples{
\dontrun{
click_pixels(3, 4)
}
}
