% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locator.R
\name{click_pixels}
\alias{click_pixels}
\title{Create Pixel Art, Get a Matrix}
\usage{
click_pixels(m)
}
\arguments{
\item{m}{A matrix. Can only contain values of 0 and 1.}
}
\value{
A matrix.
}
\description{
A provided matrix will be opened in the plotting window as a grid of
clickable squares. These can be clicked between 'on' and 'off' states.
Press ESCAPE to exit the interactive mode and be returned a matrix with the
updated values.
}
\examples{
\dontrun{
m <- matrix(sample(c(0, 1), 12, replace = TRUE), nrow = 3, ncol = 4)
click_pixels(m)
}
}
