% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/click.R
\name{click_pixels}
\alias{click_pixels}
\title{Click 'Pixels' in an Interactive Plot}
\usage{
click_pixels(
  n_rows = 8L,
  n_cols = 8L,
  n_states = 2L,
  colours = NULL,
  grid = TRUE
)
}
\arguments{
\item{n_rows}{Integer. The number of 'pixels' high that the plot should be.
Numeric values are coerced to integer.}

\item{n_cols}{Integer. The number of 'pixels' wide that the plot should be.
Numeric values are coerced to integer.}

\item{n_states}{Integer. The number of states that a pixel can be. Click a
pixel to cycle through the states. Numeric values are coerced to integer.
See details.}

\item{colours}{Character vector. As many named/hex colours as n_state. The
provided order is the order that colours will be cycled through when
pixels are clicked.}

\item{grid}{Logical. Should a boundary line be placed around the pixels to
make them easier to differentiate? Defaults to TRUE.}
}
\value{
A matrix with two attributes: 'n_states' is the number of pixel
state values provided to the function, and 'colours', which is the
character vector of colours provided to the function by the user, or, if
NULL, then a gradated set of greys provided by default.
}
\description{
Opens a new interactive plotting canvas with a grid of clickable squares
('pixels').
}
\details{
Click repeatedly the pixels in the interactive plotting window to
cycle through a number of 'states'. The initial state value is 0 and
successive clicks increase it by 1, wrapping back to 0 after the maximum
state value has been reached. Press the ESCAPE key to exit the
interactive mode and be returned a matrix that contains the state value
of each pixel.
}
\examples{
\dontrun{
    # Create a 16 x 16 pixel matrix with 3 possible pixel states
    my_matrix <- click_pixels(n_rows = 16, n_cols = 16, n_states = 3)
}
}
