% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/click.R
\name{click_pixels}
\alias{click_pixels}
\title{Click 'Pixels' in an Interactive Plot}
\usage{
click_pixels(n_rows = 8L, n_cols = 8L, n_states = 2L, grid = TRUE)
}
\arguments{
\item{n_rows}{Integer. The number of 'pixels' high that the plot should be.
Numeric values are coerced to integer.}

\item{n_cols}{Integer. The number of 'pixels' wide that the plot should be.
Numeric values are coerced to integer.}

\item{n_states}{Integer. The number of states that a pixel can be. Click a
pixel to cycle through the states. Numeric values are coerced to integer.
See details.}

\item{grid}{Logical. Should a boundary line be placed around the pixels to
make them easier to differentiate? Defaults to TRUE.}
}
\value{
A matrix.
}
\description{
Opens a new interactive plotting canvas with a grid of clickable squares
('pixels').
}
\details{
Click the pixels in the plotting window repeatedly to cycle through
a number of 'states'. Successive clicks increase the state value by 1
(wrapping back to 0, the default when the canvas is first plotted) and
make the pixel darker grey in colour. Press the ESCAPE key to exit the
mode and be returned a matrix that contains the state values of each
pixel.
}
\examples{
\dontrun{
# Create a 16 x 16 pixel matrix with 3 possible states to cycle through
click_pixels(n_rows = 16, n_cols = 16, n_states = 3, grid = TRUE)
}
}
