% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_pixels}
\alias{draw_pixels}
\title{Draw 'Pixel' Matrix in a Plot Window}
\usage{
draw_pixels(m, colours = NULL)
}
\arguments{
\item{m}{A matrix of integers. The maximum value is assumed to be the number
of pixel states desired. Override by supplying a 'n_states' value larger
than the maximum in the matrix.}

\item{colours}{Character vector. As many named/hex colours as unique state
values in 'm'. Each click in the interactive plot will cycle a pixel
through these colours. Defaults to NULL, which results in an attempt to
find and use the 'colours' attribute (a named character vector) of the
input matrix, 'm', which is added by default to matrices created
with \code{\link{click_pixels}} (recommended).}
}
\value{
Nothing.
}
\description{
Opens a plot window and draws a provided matrix. Designed for use with the
output from \code{\link{click_pixels}} (or \code{\link{edit_pixels}}).
}
\examples{
\dontrun{
  my_matrix <- click_pixels(n_states = 3L)
  draw_pixels(my_matrix, c("black", "#0000FF", "green"))  # one colour per state
}
}
