% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/click.R
\name{edit_pixels}
\alias{edit_pixels}
\title{Edit 'Pixels' in an Interactive Plot}
\usage{
edit_pixels(m, n_states = NULL, colours = NULL, grid = TRUE)
}
\arguments{
\item{m}{A matrix of integers. The maximum value is assumed to be the number
of pixel states desired. Override by supplying a 'n_states' value larger
than the maximum in the matrix.}

\item{n_states}{Integer. The number of states that a pixel can be cycled
through with successive clicks. Numeric values are coerced to integer.
Defaults to NULL, which results in an attempt to find and use the
'n_states' attribute (integer) of the input matrix, 'm'. The attribute is
added by default to matrices created with \code{\link{click_pixels}}
(recommended).}

\item{colours}{Character vector. As many named/hex colours as 'n_state'. Each
click in the interactive plot will cycle a pixel through these colours.
Defaults to NULL, which results in an attempt to find and use the
'colours' attribute (a named character vector) of the input matrix, 'm'.
This attribute is added by default to matrices created with
\code{\link{click_pixels}} (recommended).}

\item{grid}{Logical. Should a black boundary line be placed around the pixels
to help differentiate between them? Defaults to TRUE.}
}
\value{
A matrix.
}
\description{
Opens an interactive plotting canvas with a grid of clickable squares
('pixels') that represent the cells of a matrix provided by the user,
ideally the output from \code{\link{click_pixels}}.
}
\details{
Click repeatedly the pixels in the interactive plotting window to
cycle through the provided number of 'states'. The initial state value is
0 and successive clicks increase it by 1, wrapping back to 0 once the
maximum number of states is exceeded. Press the ESCAPE key to exit the
interactive mode.

If your editor opens a separate graphics window (i.e. not RStudio),
each click may result in a brief flash as the image refreshes, while a
resized window may return to its original dimensions. You may also hear a
bell sound on click, which you can disable by setting
\code{options(locatorBell = FALSE)}.
}
\examples{
\dontrun{
    # Create a 3 x 4 pixel matrix with 3 possible states to cycle through
    my_matrix <- click_pixels(
      n_rows   = 3L,
      n_cols   = 4L,
      n_states = 3L,
      colours  = c("white", "red", "#0000FF")
    )

.    # Update the original matrix
    my_matrix_edited <- edit_pixels(m = my_matrix)

    # Update the original matrix with additional state, different colours
    my_matrix_augmented <- edit_pixels(
      m        = my_matrix,
      n_states = 4L,  # one more than in the original
      colours  = c("bisque3", "orchid", "chartreuse", "olivedrab")
    )
}
}
