% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit.R
\name{edit_pixels}
\alias{edit_pixels}
\title{Edit 'Pixels' in an Interactive Plot}
\usage{
edit_pixels(m, n_states = NULL, grid = TRUE)
}
\arguments{
\item{m}{A matrix of integers. The maximum value is assumed to be the number
of pixel states desired. Override by supplying a 'n_states' value larger
than the maximum in the matrix.}

\item{n_states}{Integer. The number of states that a pixel can be. Click a
pixel to cycle through the states. Numeric values are coerced to integer.
See details.}

\item{grid}{Logical. Should a boundary line be placed around the pixels to
make them easier to differentiate? Defaults to TRUE.}
}
\value{
A matrix.
}
\description{
Opens an interactive plotting canvas with a grid of clickable squares
('pixels') that represent the cells of a matrix provided by the user,
ideally the output from \code{\link{click_pixels}}.
}
\details{
Click the pixels in the plotting window repeatedly to cycle through
a number of 'states'. Successive clicks increase the state value by 1
(wrapping back to 0, the default when the canvas is first plotted) and
make the pixel darker grey in colour. Press the ESCAPE key to exit the
mode and be returned a matrix that contains the state values of each
pixel.
}
\examples{
\dontrun{edit_pixels(m, 3, FALSE)}
}
