% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{frame_pixels}
\alias{frame_pixels}
\title{Create Frames of a Pixel Animation}
\usage{
frame_pixels(...)
}
\arguments{
\item{...}{Parameters to pass to \code{\link{click_pixels}}.}
}
\value{
A list of matrices.
}
\description{
Opens a new interactive plotting canvas with a grid of clickable squares
('pixels'). When finished, the user is prompted to provide another.
}
\examples{
\dontrun{m <- frame_pixels(n_rows = 8, n_cols = 8, n_states = 3)}
}
