% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{gif_pixels}
\alias{gif_pixels}
\title{Write Frames of a Pixel Animation to GIF}
\usage{
gif_pixels(frames, colours = NULL, file, ...)
}
\arguments{
\item{frames}{A list of matrices preferably produced by \code{\link{frame_pixels}}.
Each is a matrix with the same dimension and possible values that
represents a single frame of an animation.}

\item{colours}{A character vector of named colours. One for each value in m,
the provided matrix. The order you provide the colours matches the order
of the values in the matrix, from 0 to n. Defaults to NULL, which means
random colours will be selected on your behalf.}

\item{file}{A filepath expressed as a character vector, ending with
extension '.gif'. This is where the output animation will be written.}

\item{...}{Parameters to pass to \code{\link[gifski]{save_gif}} (you must
have 'gifski' installed).}
}
\value{
Nothing.
}
\description{
Draws each list element (matrices) from the output of \code{\link{frame_pixels}}
and writes it to an animated gif. Requires user to install the 'gifski'
package.
}
\examples{
\dontrun{
    # Create interactively a list of matrices, i.e. frames of an animation
    my_frames <- frame_pixels(n_rows = 8, n_cols = 8, n_states = 3)

    # Write
    gif_pixels(
      frames = my_frames,
      colours = c("black", "blue", "#0000FF"),  # one colour per matrix value
      tempfile("example", ".gif")  # location to write GIF
    )
}
}
