% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixeltrix-package.R
\docType{package}
\name{pixeltrix-package}
\alias{pixeltrix}
\alias{pixeltrix-package}
\title{pixeltrix: Convert Interactive Pixel Art to a Matrix}
\description{
An interactive 'app' where the user interactively clicks 'pixels' in an image and it's converted into a matrix object.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matt-dray/pixeltrix}
  \item Report bugs at \url{https://github.com/matt-dray/pixeltrix/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
