% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixeltrix-package.R
\docType{package}
\name{pixeltrix-package}
\alias{pixeltrix}
\alias{pixeltrix-package}
\title{pixeltrix: Simple Interactive Pixel Art}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A very simple 'pixel art' tool that lets you click squares interactively in a plot window and returns your final image as a matrix. Provide multiple frames to create an animated gif.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matt-dray/pixeltrix}
  \item \url{https://matt-dray.github.io/pixeltrix}
  \item Report bugs at \url{https://github.com/matt-dray/pixeltrix/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
