% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixeltrix-package.R
\docType{package}
\name{pixeltrix-package}
\alias{pixeltrix}
\alias{pixeltrix-package}
\title{pixeltrix: Create Matrices from Users' Pixel Art}
\description{
A Shiny app where the user interactively draws a 'pixel art' image and it gets converted into a copyable matrix object.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matt-dray/pixeltrix}
  \item Report bugs at \url{https://github.com/matt-dray/pixeltrix/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
