% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixeltrix-package.R
\docType{package}
\name{pixeltrix-package}
\alias{pixeltrix}
\alias{pixeltrix-package}
\title{pixeltrix: Click Pixels and Get a Matrix}
\description{
A very simple tool that lets you click ‘pixels’ interactively in a plot window and returns your final image as a matrix. Could be used to help create simple videogame sprites, for example.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matt-dray/pixeltrix}
  \item \url{https://matt-dray.github.io/pixeltrix}
  \item Report bugs at \url{https://github.com/matt-dray/pixeltrix/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
